init the popup opacity slider
svn path=/trunk/koffice/; revision=662282

+++ krita/ui/kis_layerbox.h
@@ -67,7 +67,7 @@ public:
 
 public slots:
     void slotSetCompositeOp(const KoCompositeOp* compositeOp);
-    void slotSetOpacity(int opacity);
+    void slotSetOpacity(double opacity);
     void slotSetColorSpace(const KoColorSpace * colorSpace);
 
 signals:
